Move PropertyType to org.eclipse.ec4j.core.model because it is a model class
