+++ core/src/main/java/org/eclipse/ec4j/core/PropertyTypeRegistry.java
@@ -23,7 +23,7 @@ import java.util.Locale;
 import java.util.Map;
 
 import org.eclipse.ec4j.core.model.Property;
-import org.eclipse.ec4j.core.model.propertytype.PropertyType;
+import org.eclipse.ec4j.core.model.PropertyType;
 
 /**
  * A mapping from property names to {@link PropertyType}s. Note that the mapping is case insensitive - i.e. all names
