+++ core/src/main/java/org/eclipse/ec4j/core/model/Section.java
@@ -24,7 +24,6 @@ import java.util.List;
 import java.util.Map;
 
 import org.eclipse.ec4j.core.model.propertytype.IndentStyleValue;
-import org.eclipse.ec4j.core.model.propertytype.PropertyType;
 
 /**
  * A section in an {@code .editorconfig} file. A section consists of a {@link Glob} and a collection of
