+++ services/src/main/java/org/eclipse/ec4j/services/EditorConfigService.java
@@ -23,10 +23,10 @@ import java.util.List;
 import java.util.Set;
 
 import org.eclipse.ec4j.core.EditorConfigConstants;
+import org.eclipse.ec4j.core.PropertyTypeRegistry;
 import org.eclipse.ec4j.core.Resources;
 import org.eclipse.ec4j.core.Resources.RandomReader;
-import org.eclipse.ec4j.core.model.propertytype.PropertyType;
-import org.eclipse.ec4j.core.model.propertytype.PropertyTypeRegistry;
+import org.eclipse.ec4j.core.model.PropertyType;
 import org.eclipse.ec4j.core.parser.EditorConfigParser;
 import org.eclipse.ec4j.services.completion.CompletionContextType;
 import org.eclipse.ec4j.services.completion.CompletionEntry;
